adjustments to texts, and move feedback component to its own file
